move tenant middlewware to separate file and make it exported to allow tests to pass
